pyproject.toml
[project]
name = "infinite-realms"
version = "1.0.0"
description = "AI-powered D&D 5E Campaign Bible Generator using CrewAI"
authors = [{name = "Infinite Realms Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.98.0",
    "crewai-tools>=0.15.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "chromadb>=0.4.0",
    "markdown>=3.5.0",
    "jinja2>=3.1.0",
    "networkx>=3.2.0",
    "matplotlib>=3.8.0",
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
